+++ user.h
@@ -10,8 +10,8 @@ typedef struct user_t {
 	bool public;
 } user_t;
 
-int find_user_by_id(sqlite3 *database, uint8_t (*user_id)[16], user_t *user);
-int find_user_by_name(sqlite3 *database, char *name, size_t name_len, user_t *user);
+uint16_t find_user_by_id(sqlite3 *database, uint8_t (*user_id)[16], user_t *user);
+uint16_t find_user_by_name(sqlite3 *database, char *name, size_t name_len, user_t *user);
 
 void find_user(sqlite3 *database, bwt_t *bwt, response_t *response);
 void delete_user(sqlite3 *database, bwt_t *bwt, response_t *response);
